backends/binance/mcp_server/requirements.txt
# Core dependencies
mcp>=1.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=0.19.0
PyYAML>=6.0
httpx>=0.24.0

# MongoDB dependencies
motor>=3.3.0
pymongo>=4.6.0

# Development tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Type checking
typing_extensions>=4.0.0

# Testing
pytest-mock>=3.10.0
asynctest>=0.13.0

# Additional utilities
statistics>=1.0.3.5  # For mean calculation
python-dateutil>=2.8.2  # For date handling
